Variable cleanup and modifications to build in eclipse

+++ ConfigureNyboard/Globals.h
@@ -54,7 +54,7 @@ class Globals {
     static Command curr_command;
     static Command last_command;
 
-    static int16_t command_vals[MAX_COMMAND_VALUES];
+    static int command_vals[MAX_COMMAND_VALUES];
 
     static Motion motion;
